Remove extra wrapping for model
